+++ packages/components-react/src/ModalDialog/index.ts
@@ -1,0 +1,14 @@
+export {
+  ModalDialog,
+  ModalDialogHeader,
+  ModalDialogHeading,
+  ModalDialogBody,
+  ModalDialogFooter,
+} from './ModalDialog';
+export type {
+  ModalDialogProps,
+  ModalDialogHeaderProps,
+  ModalDialogHeadingProps,
+  ModalDialogBodyProps,
+  ModalDialogFooterProps,
+} from './ModalDialog';