+++ packages/components-react/src/index.ts
@@ -70,3 +70,4 @@ export * from './FormFieldLegend';
 
 export * from './Image';
 export * from './Card';
+export * from './ModalDialog';